initial multiple typescript files
