+++ packages/modelviewer.dev/examples/examples-and-docs.html
@@ -255,7 +255,7 @@
     <div id="loading-examples"></div>
   </div>
 
-  <script type="module" src="./built/chris-example.js"></script>
+  <script type="module" src="./built/documentation-main.js"></script>
 
   <!-- Documentation-specific dependencies: -->
   <script type="module" src="./built/dependencies.js">
